change file to tsx, add component props
